+++ src/components/ContentOrderedListRenderer.svelte
@@ -40,10 +40,6 @@
     counter-increment: item;
   }
 
-  li > :global(*) {
-    background: var(--list-background);
-  }
-
   ol.phr-reversed li {
     counter-increment: item -1;
   }