+++ src/components/ContentRenderer.svelte
@@ -50,6 +50,12 @@
 {/if}
 
 <style>
+  span {
+    background: var(--background);
+    /* text would probably be better, but browser support is still bad */
+    background-clip: content-box;
+  }
+
   hr {
     margin: 24pt 0;
   }