+++ src/components/ContentUnorderedListRenderer.svelte
@@ -30,10 +30,6 @@
     padding-left: 1em;
   }
 
-  li > :global(*) {
-    background: var(--list-background);
-  }
-
   li::marker {
     color: var(--marker-color);
   }